antispoof/requirements.txt
# antispoof — pinned dependencies
# macOS M-series: install dlib with: pip install dlib --no-cache-dir
# Kaggle/Colab: cmake is pre-installed

torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.0.200        # YOLOv8
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
dlib>=19.24.2               # needs cmake: brew install cmake / apt install cmake libopenblas-dev
onnx>=1.15.0
onnxruntime>=1.16.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6     # FastAPI file upload
slowapi>=0.1.9              # rate limiting
pydantic>=2.4.0
roboflow>=1.1.9
gdown>=4.7.3
tqdm>=4.66.0
tensorboard>=2.14.0
seaborn>=0.13.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
